Untangle process transition recording types
