+++ crates/voom-cli/src/commands/workflow/process/transitions.rs
@@ -3,11 +3,7 @@
 //! Collects the information needed to build a `FileTransition` without
 //! requiring callers to pass the whole process orchestration context.
 
-/// Dependencies needed to persist file transitions.
-pub(super) struct TransitionRecorder<'a> {
-    pub store: &'a dyn voom_domain::storage::StorageTrait,
-    pub session_id: uuid::Uuid,
-}
+use super::context::TransitionRecorder;
 
 /// Grouped arguments for `record_file_transition`.
 pub(super) struct FileTransitionContext<'a> {
@@ -78,68 +74,11 @@ pub(super) fn record_file_transition(tctx: &FileTransitionContext<'_>) {
     }
 }
 
-/// Grouped arguments for `record_failure_transition`.
-///
-/// The `executor` should be the executor plugin name, or an empty string when
-/// no executor was involved (e.g. safeguard abort).
-pub(super) struct FailureTransitionContext<'a> {
-    pub file: &'a voom_domain::media::MediaFile,
-    pub plan: &'a voom_domain::plan::Plan,
-    pub executor: &'a str,
-    pub error_message: Option<&'a str>,
-    pub recorder: &'a TransitionRecorder<'a>,
-}
-
-/// Record a failure transition in the store for a plan that did not succeed.
-///
-/// The file is unchanged on failure, so `to_size = from_size` and `to_hash =
-/// from_hash`.
-///
-/// Dual-write: `file_transitions.error_message` is used for session-based
-/// queries (`voom report errors`), while `plans.result` stores the structured
-/// `ExecutionDetail` JSON for plan-based queries with full subprocess output.
-pub(super) fn record_failure_transition(fctx: &FailureTransitionContext<'_>) {
-    let file = fctx.file;
-    let to_hash = file.content_hash.clone().unwrap_or_default();
-    let mut transition = voom_domain::FileTransition::new(
-        file.id,
-        file.path.clone(),
-        to_hash,
-        file.size,
-        voom_domain::TransitionSource::Voom,
-    )
-    .with_from(file.content_hash.clone(), Some(file.size))
-    .with_detail(fctx.executor)
-    .with_plan_id(fctx.plan.id)
-    .with_processing(
-        0,
-        0,
-        0,
-        voom_domain::ProcessingOutcome::Failure,
-        &fctx.plan.policy_name,
-        &fctx.plan.phase_name,
-    )
-    .with_session_id(fctx.recorder.session_id);
-
-    if let Some(msg) = fctx.error_message {
-        transition = transition.with_error_message(msg);
-    }
-
-    if let Err(e) = fctx.recorder.store.record_transition(&transition) {
-        tracing::warn!(
-            path = %fctx.file.path.display(),
-            error = %e,
-            "failed to record failure transition"
-        );
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::path::PathBuf;
 
     use voom_domain::media::MediaFile;
-    use voom_domain::plan::Plan;
     use voom_domain::storage::FileTransitionStorage;
     use voom_domain::test_support::InMemoryStore;
 
@@ -223,33 +162,5 @@ mod tests {
         assert_eq!(transition.duration_ms, Some(25));
         assert_eq!(transition.actions_taken, Some(2));
         assert_eq!(transition.tracks_modified, Some(1));
-    }
-
-    #[test]
-    fn record_failure_transition_persists_error_message() {
-        let store = InMemoryStore::new();
-        let file = media_file("/movies/example.mkv", "hash-a");
-        let plan = Plan::new(file.clone(), "policy", "metadata");
-        let recorder = TransitionRecorder {
-            store: &store,
-            session_id: uuid::Uuid::new_v4(),
-        };
-
-        record_failure_transition(&FailureTransitionContext {
-            file: &file,
-            plan: &plan,
-            executor: "ffmpeg-executor",
-            error_message: Some("bad codec"),
-            recorder: &recorder,
-        });
-
-        let transitions = store
-            .transitions_for_file(&file.id)
-            .expect("transitions query");
-        assert_eq!(transitions.len(), 1);
-        let transition = &transitions[0];
-        assert_eq!(transition.error_message.as_deref(), Some("bad codec"));
-        assert_eq!(transition.source_detail.as_deref(), Some("ffmpeg-executor"));
-        assert_eq!(transition.plan_id, Some(plan.id));
     }
 }
